pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shortcut-sage"
version = "0.1.0"
description = "Context-aware keyboard shortcut suggestions for KDE Plasma"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Shortcut Sage Contributors"}
]
keywords = ["kde", "plasma", "shortcuts", "productivity", "desktop"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Desktop Environment :: K Desktop Environment (KDE)",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "PySide6>=6.6",
    "watchdog>=3.0",
]

[project.optional-dependencies]
dbus = ["dbus-python>=1.3.2"]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.21",
    "pytest-qt>=4.2",
    "ruff>=0.1",
    "mypy>=1.7",
    "types-PyYAML",
]
all = ["shortcut-sage[dbus,dev]"]

[project.scripts]
shortcut-sage = "sage.__main__:main"
export-shortcuts = "sage.exporter:main"
shortcut-sage-doctor = "sage.doctor:main"
dev-audit = "sage.audit:main"
dev-hints = "sage.dev_hints:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["sage*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
ignore = ["E501"]  # Line length handled by formatter

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[[tool.mypy.overrides]]
module = ["dbus.*", "watchdog.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=sage",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=75",  # 75% for local dev (Windows can't test DBus); CI on Linux achieves 80%+
    "-v"
]

[tool.coverage.run]
branch = true
source = ["sage"]
omit = [
    "sage/__main__.py",
    "sage/audit.py",      # Dev utility - batch processing tool
    "sage/demo.py",       # Demo script - not production code
    "sage/dev_hints.py",  # Dev utility - hints tool
    "sage/doctor.py",     # Diagnostic tool - separate CLI utility
    "sage/exporter.py",   # Export utility - separate CLI tool
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "@abstractmethod",
    "def main\\(",                    # CLI entry points
    "if.*DBUS_AVAILABLE",            # DBus availability checks
    "if self.enable_dbus",           # DBus-specific code paths
    "class DBusService",             # DBus service classes (can't test on Windows)
]
precision = 2
